Chore: update unit-tests script
Using the QtWebEngineWidgets requires the use of MSVC and qt5. However,
running from MSVC requires that the x64 Native Tools environment
variables are set. So in addition to updating the qt version to 5, I
added a note explaining the previous statement. If this script is ran
without the environment variables set properly, the cmake build will
fail.