Feat: add basic POESESSID capture support
Whenever you log in to the official POE website, the POESESSID cookie is
generated. This cookie can be used to by other tools to access your
profile/stash tabs/etc. This update saves the POESESSID whenever it is
created, so that the user can easily extract and copy it for other tools
if necessary.