Chore: update pre-commit hook to not cancel commit
Commiting should be done often, and changes to be committed do not
always leave the application in a working state. Previously, the
pre-commit hook would refuse to commit if the static analysis or unit
tests failed. This was changed to still run those checks to inform the
developer about any issues, but still allow the commit to go through