Feat(WIP): add GetLatestReleaseUrl func
This func is used to translate a github.com/user/repo path to
api.github.com/repos/user/repo/releases/latest. This url will be used to
get download the latest release of the repository. Right now it only
validates the path by seeing if it contains 'github.com'. In theory, a
regex could be used to make sure it fits the expected input, but right
now it will just fail the http request.