Refactor: reduce scope of AppManager class
My initial idea of the AddApp() func was that if given a GitHub URL, it
the Manager would go through the process of downloading the app from
GitHub. But because a GitHub based app is very complicatd (what if there
are multiple asset files/executables?) if made the AppManager class
riddled with GitHub related functions. I decided to remove those from
the AppManager class and will make some sort of ApiHandler class for
those later. As of now, the AppManager's scope is to deal with the
collection of Apps: adding, removing, getting a specific app from list,
etc.