Feat(WIP): improve getting latest repo release
Improved the method for getting the latest release of a repository.
Because the only real reason to query for the latest release is to
potentially download it, the assets of the release are now bundled with
the releases name.

Also added a helper function to retrieve the data from a QNetworkReply.
This will be useful when I add more Api functions.