Feat(WIP): improve downloading of github assets
Added ability to download release asset from GitHub and extract it to a
folder (currently ./apps/).
Added ability to choose between multiple assets, if found.
Added ability to extract the repo information from a github url

It is now possible to give ExileHub a github URL and it will download
the asset and add the executable to the AppMananger to run.

Will need to add ability to choose between multiple if executables if
found.