Feat(WIP): add ability to activate app windows
Added the ability to open windows of apps that all already running. This
will allow users to maximize/show windows from within exilehub and not
have to worry about finding it on the taskbar. This will also be used
when hotkeys are implemented, allowing the user to open the app with
hotkeys.