Feat(WIP): add auto-start and auto-update feats
Added ability to start/check for updates for a program when exilehub
opens. Eventually, it would be preferable to have the apps open/close
with PoE itself. Currently, the update prompt shows before the main
window, and doesn't indicate which app needs updating. This will be
fixed in the future.