Fix: fix issue with autoupdate and autolaunch
On start, apps are checked for updates before auto-launching. Because
the user is asked after updating if they want to start the program,
there was an issue where they could choose no, but the app started
anyways because of the auto-launching. This change makes it so that if
the user was prompted about launching the app, that auto-launching is
ignored